with validation

+++ Task2-Radency-Internship/Controllers/BookController.cs
@@ -18,11 +18,14 @@ namespace Task2_Radency_Internship.Controllers
             _config = config;
         }
         [HttpGet]
-        [Route("[action]")]
-        public async Task<ActionResult<List<BookWithRatingDto>>> Books(/*[FromQuery] */ string? order)
+        [Route("books")]
+        public async Task<ActionResult<List<BookWithRatingDto>>> GetAllBooks(/*[FromQuery] */ string? order)
         {
             var books = await _bookService.GetAllBooksAsync();
-
+            if (books.Count == 0)
+            {
+                return BadRequest("Library is empty");
+            }
             var result = books.Select(book => BookWithRatingDto.ToDtoModel(book)).ToList();
             if (order == "author")
             {
@@ -43,6 +46,10 @@ namespace Task2_Radency_Internship.Controllers
         public async Task<ActionResult<List<BookWithRatingDto>>> Recommended(string? genre)
         {
             var books = await _bookService.GetAllBooksAsync();
+            if(books.Count == 0)
+            {
+                return BadRequest("Library is empty");
+            }
             var booksWithRating = books.Select(book => BookWithRatingDto.ToDtoModel(book)).ToList();
             var result = booksWithRating.OrderByDescending(book => book.Ratings).Where(p => p.ReviewsNumber > 2).Take(10).ToList();   //For testing purpose, required ReviewsNumber reduced to 2
             if (genre != null)
@@ -53,29 +60,40 @@ namespace Task2_Radency_Internship.Controllers
             return Ok(result);
         }
         [HttpGet]
-        [Route("[action]/{id}")]
-        public async Task<ActionResult<List<BookDetailedReviewDto>>> Books(int id)
+        [Route("books/{id}")]
+        public async Task<ActionResult<List<BookDetailedReviewDto>>> GetBook(int id)
         {
             var books = await _bookService.GetBookAsync(id);
-            var result = BookDetailedReviewDto.ToDtoModel(books);
-            return Ok(result);
-
+            if(books != null)
+            {
+                var result = BookDetailedReviewDto.ToDtoModel(books);
+                return Ok(result);
+            }
+            return BadRequest($"Book with id: {id} did'nt exist ");
         }
         [HttpDelete]
-        [Route("[action]/{id}")]
-        public async Task<ActionResult> Books(int id, string? secret)
+        [Route("books/{id}")]
+        public async Task<ActionResult> DeleteBook(int id, string? secret)
         {
-            if (secret == _config["SecretWord"])
+            var bookexist = await _bookService.GetBookAsync(id);
+            if (bookexist != null)
             {
-                await _bookService.DeleteBook(id);
+
+                if (secret == _config["SecretWord"])
+                {
+                    await _bookService.DeleteBook(id);
+                }
+                else
+                {
+                   return BadRequest("Incorrect secret word");
+                }
             }
             return Accepted();
         }
         [HttpPost]
-        [Route("[action]/save")]
-        public async Task<ActionResult> Books([FromBody] NewBookDto newBookDto)
+        [Route("books/save")]
+        public async Task<ActionResult> CreateOrUpdate([FromBody] NewBookDto newBookDto)
         {
-
             var book = newBookDto.ToModel();
             var bookexist = await _bookService.GetBookAsync(newBookDto.Id);
             if (bookexist != null)
@@ -99,6 +117,7 @@ namespace Task2_Radency_Internship.Controllers
         [Route("books/{id}/[action]")]
         public async Task<ActionResult> Review(int id, [FromBody] ReviewDto reviewDto)
         {
+      
             var bookexist = await _bookService.GetBookAsync(id);
             if (bookexist != null)
             {
@@ -106,29 +125,37 @@ namespace Task2_Radency_Internship.Controllers
                 review.BookId = id;
                 bookexist.Reviews.Add(review);
                 await _bookService.SaveChanges();
+                return Accepted(new { id = bookexist.Reviews.Last().Id });
             }
-            return Accepted(new { id = bookexist.Reviews.Last().Id });
+            return BadRequest($"Book with id: {id} did'nt exist ");
+            
         }
         [HttpPut]
         [Route("books/{id}/[action]")]
         public async Task<ActionResult> rate(int id, [FromBody] RatingDto ratingDto)
         {
+
             var bookexist = await _bookService.GetBookAsync(id);
-            if (ratingDto.Score >= 1 && ratingDto.Score <= 5)
+            if (bookexist != null)
             {
-                
-                if (bookexist != null)
+
+                if (ratingDto.Score >= 1 && ratingDto.Score <= 5)
                 {
                     var rate = ratingDto.ToModel();
                     rate.BookId = id;
                     bookexist.Ratings.Add(rate);
                     await _bookService.SaveChanges();
                 }
+                else
+                {
+                    return BadRequest("score must be between dfgdfgdfg1 and 5");
+                }
             }
             else
             {
-                return BadRequest("score must be between 1 and 5");
+                return BadRequest($"Book with id: {id} did'nt exist ");
             }
+
 
             return Accepted(new { id = bookexist.Ratings.Last().Id });
         }
